tests/wpt/tests/navigation-api/navigation-methods/return-value/navigate-rejection-order-pagehide-unserializablestate.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/helpers.js"></script>

<iframe id="i" src="/common/blank.html"></iframe>

<script>
promise_test(async t => {
  await new Promise(resolve => window.onload = resolve);

  let navigateEventCount = 0;
  i.contentWindow.navigation.onnavigate = () => navigateEventCount++;
  i.contentWindow.navigation.onnavigatesuccess = t.unreached_func("onnavigatesuccess should not be called");
  i.contentWindow.navigation.onnavigateerror = t.unreached_func("onnavigateerror should not be called");

  i.contentWindow.navigation.navigate("?1");

  await new Promise(resolve => {
    i.contentWindow.onpagehide = t.step_func(async () => {
      await assertBothRejectDOM(t, i.contentWindow.navigation.navigate("?2", { state: document.body }), "DataCloneError", i.contentWindow);
      assert_equals(navigateEventCount, 1);
      resolve();
    });
  });
}, `navigate() with an unserializable state inside onpagehide throws "DataCloneError", not "InvalidStateError"`);
</script>
