LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-methods/return-value/navigate-rejection-order-pagehide-unserializablestate-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_equals: expected 1 but got 0


Harness Error (FAIL), message = Unhandled rejection: assert_equals: expected 1 but got 0

TIMEOUT navigate() with an unserializable state inside onpagehide throws "DataCloneError", not "InvalidStateError" Test timed out

Harness Error (FAIL), message = Unhandled rejection: assert_equals: expected 1 but got 0

TIMEOUT navigate() with an unserializable state inside onpagehide throws "DataCloneError", not "InvalidStateError" Test timed out
